+++ plugin/controllers/web.py
@@ -14,7 +14,7 @@ from Plugins.Extensions.OpenWebif.__init__ import _
 from Components.config import config
 
 from models.info import getInfo, getCurrentTime , getStatusInfo, getFrontendStatus
-from models.services import getCurrentService, getBouquets, getServices, getSubServices, getChannels, getSatellites, getBouquetEpg, getBouquetNowNextEpg, getSearchEpg, getChannelEpg, getNowNextEpg, getSearchSimilarEpg, getAllServices, getPlayableServices, getPlayableService, getParentalControlList
+from models.services import getCurrentService, getBouquets, getServices, getSubServices, getChannels, getSatellites, getBouquetEpg, getBouquetNowNextEpg, getSearchEpg, getChannelEpg, getNowNextEpg, getSearchSimilarEpg, getAllServices, getPlayableServices, getPlayableService, getParentalControlList, getEvent
 from models.volume import getVolumeStatus, setVolumeUp, setVolumeDown, setVolumeMute, setVolume
 from models.audiotrack import getAudioTracks, setAudioTrack
 from models.control import zapService, remoteControl, setPowerState, getStandbyState
@@ -330,7 +330,13 @@ class WebController(BaseController):
 		return renameMovie(self.session, request.args["sRef"][0],request.args["newname"][0])
 
 	def P_movietags(self, request):
-		return getMovieTags()
+		_add = None
+		_del = None
+		if "add" in request.args.keys():
+			_add = request.args["add"][0]
+		if "del" in request.args.keys():
+			_del = request.args["del"][0]
+		return getMovieTags(_add,_del)
 
 	# a duplicate api ??
 	def P_gettags(self, request):
@@ -707,6 +713,12 @@ class WebController(BaseController):
 
 		return getSearchSimilarEpg(request.args["sRef"][0], eventid)
 
+	def P_event(self, request):
+		event = getEvent(request.args["sref"][0], request.args["idev"][0])
+		event['event']['recording_margin_before'] = config.recording.margin_before.value
+		event['event']['recording_margin_after'] = config.recording.margin_after.value
+		return event
+	
 	def P_getcurrent(self, request):
 		info = getCurrentService(self.session)
 		now = getNowNextEpg(info["ref"], 0)